Implemented the string[] into the nodebasedata
I then adjusted the saving method to save string arrays and not singular string
loading as well (connection stitching part)
Just need to now write the switch case node and interact with the outpoint having an array of out points.

Then adjust the code as welll to that change and i can be good to go for to do the remove unused events button code

Also find a way to make it a MUST to implement a function called Unpack and Setup in the LEM_Effects scripts. This way i wont forget