pt 4 Redo and undo for Delete and Create Node commands done!!
Fixed deadly issue where AllNodesInEditor couldnt remove a node even thou there is a node with the same NodeID inside but due to Remove() not checking the NodeID but rather the reference instead, the node cannot be removed. But i fixed that and made the removal process still the same Big O Notation as the old method.