pt 7 Standard of Life update 2
 Cut paste, cut paste ,cut paste, undo undo undo undo undo = connections not removed <Priority> FXIED

Fixed it but i need to convert the function of CompiletoBaseEffect to accept nodeid. Inother words, PHASE OUT USE OF BASEEFFECTNODE[ ] in all commands.

After shallowcloning the stitching part of the connections had issues. FIXEDCAUSE I THOUGHT STRING IS A VALUE TYPE BUT TURNS OUT IT WAS A REFERENCE TYPE FUCK YOU AHAHAHHAHA